feat: SkillNameList 컴포넌트 분리
